+++ src/Pan.h
@@ -48,7 +48,7 @@ public:
    ~PanChannel() override;
 
 protected:
-   double update(double time) override;
+   double update(uint64_t time) override;
 
 private:
    Pan *_host;
